Moved user information to auth
